+++ src/Models/Addon.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Liberu\Billing\Catalog\Models;
+
+final class Addon extends CatalogRecord
+{
+    protected $table = 'billing_catalog_addons';
+}